See the css that I have started
I started learning about all sorts of css in the odin project.
I have also learned that I should commit regularly and this
may help to land me a job.

So I added some css and plan to commit daily to my git hub. I
know this commit message has little to do with what I have added
to the code but I am learning. Keeping a health record of what I'm
too and building habbits is the goal.

On the plus side, if I commit every day, it may be what lands me a
job over the guy next to me. The person looking to hire me will be
able to see that I have committed regurlarly. This may look good.