See additional flex box items
Setting up more flex boxes to prepare for the portfolio that
I am working on. Would like to start building a blank slate
for my portfolio early, that way I can keep typing and building
routines and brain cells for whats going on. This particular
project won't be my actual portfolio, just working on learning
and training myself to work this new way. On a key board, and
coding. Love it, wish me luck!