The last four boxes looking better
Learning very quickly that I need to compress the css here a
little bit. understanding that I could have alot of these rules
all in just a few css properties rather than the book I am
writing. Will be working on this in the future.

So far though, feels good to see the page start to change in
a positive way.